Move recipe to align with existing best practices recipes
