+++ src/main/java/org/openrewrite/java/recipes/CorrectlySpacedDescriptions.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.rewrite;
+package org.openrewrite.java.recipes;
 
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Preconditions;
